pyproject.toml
[tool.poetry]
name = "teamatch-backend"
version = "0.1.0"
description = ""
authors = ["BOBO-LU <luwenkai0820@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
pytz = "^2022.7.1"
gunicorn = "^20.1.0"
loguru = "^0.6.0"
pytest = "^7.2.1"
asynctest = "^0.13.0"
aiohttp = "^3.8.4"
httpx = "^0.23.3"
pytest-html = "^3.2.0"
coverage = "^7.2.1"
pytest-cov = "^4.0.0"
sqlalchemy = "^2.0.4"
fastapi-mail = "^1.2.6"
fastapi-jwt-auth = {extras = ["asymmetric"], version = "^0.5.0"}
psycopg2-binary = "^2.9.5"
passlib = "^1.7.4"
alembic = "^1.9.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
tenacity = "^8.2.2"
python-multipart = "^0.0.6"
orjson = "^3.8.9"
authlib = "^1.2.0"
itsdangerous = "^2.1.2"
pre-commit = "^3.2.2"
fastapi-sso = "^0.6.4"
google-auth = "^2.17.1"
requests = "^2.28.2"
google-api-python-client = "^2.83.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
black = "^23.3.0"
autoflake = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
  ".", "app", "backend/app"
]

[tool.autoflake]
check = true
